# api/openapi.yaml
...
openapi: 3.1.0
...
info:
  title: Hikyo API
  version: "1"
  summary: Self-hosted environment and secret management.
  description: |
    Every response may gain fields additively within `/api/v1`; clients MUST
    ignore fields they do not know. Open enums (`x-extensible-enum`) may gain
    values; closed enums never do.

    Unauthorized is indistinguishable from nonexistent on every tenant-scoped
    route: status, body and shape are identical (permission model ADR). Error
    bodies carry a fixed message per code — nothing derived from the request —
    except `bad_request`, which is decided before any tenant resolution and so
    may carry a `detail` that leaks nothing about tenancy.
  license:
    name: MPL-2.0
    identifier: MPL-2.0
servers:
  - url: /
    description: The instance this document was served from.
...
paths:
...
  /api/v1/auth/methods:
    get:
      operationId: authMethods
      tags: [auth]
      summary: Enabled login methods for this instance.
      description: |
        Public discovery: the enabled OIDC provider slugs and display names, and
        whether local login is enabled. Instance-level, never per-account, so it
        discloses nothing about which accounts or factors exist. Under the
        discovery admission bucket.
      security: []
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [none]
      x-hikyo-min-revision: 1
      responses:
        "200":
          description: The enabled login methods.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMethods"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/oidc/{provider}/start:
    parameters:
      - $ref: "#/components/parameters/ProviderSlug"
    post:
      operationId: oidcStart
      tags: [auth]
      summary: Begin an OIDC transaction (login, link or reauth).
      description: |
        Creates a single-use server-side transaction (PKCE S256 always) and
        returns the IdP authorization URL. `login` is anonymous and bound by a
        `__Host-` browser-binding cookie set on this response (A2/A16); `link`
        and `reauth` require an authenticated session and are session-bound, and
        `link` verifies the account-security proof up front (A6).
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OidcStartRequest"
      responses:
        "200":
          description: The IdP authorization URL.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcStartResult"
        # Every expected start refusal (unknown or disabled slug, bad purpose,
        # policy-less or environment-less reauth, unauthenticated link/reauth,
        # and a missing body) collapses to one uniform 401 so a pre-auth prober
        # cannot enumerate provider config. 429/500 are the cross-cutting
        # admission and fault responses.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/oidc/{provider}/callback:
    parameters:
      - $ref: "#/components/parameters/ProviderSlug"
    get:
      operationId: oidcCallback
      tags: [auth]
      summary: Complete an OIDC transaction from the IdP redirect.
      description: |
        Per-provider distinct callback path (mix-up defence leg 1): the path
        provider must equal the transaction's provider, refused BEFORE any token
        is fetched (A1). The transaction is consumed single-use, the code is
        exchanged only at the recorded provider, the ID token validated
        completely, and the purpose wall is structural. Returns the minted or
        rotated session.
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      parameters:
        - name: code
          in: query
          required: false
          schema: { type: string, maxLength: 4096 }
        - name: state
          in: query
          required: false
          schema: { type: string, maxLength: 512 }
        - name: iss
          in: query
          required: false
          schema: { type: string, maxLength: 2048 }
        - name: error
          in: query
          required: false
          schema: { type: string, maxLength: 256 }
      responses:
        "200":
          description: The minted or rotated session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/auth/identities:
    get:
      operationId: listIdentities
      tags: [auth]
      summary: List the caller's linked external identities.
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      responses:
        "200":
          description: The caller's linked identities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentityList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/identities/link:
    post:
      operationId: linkIdentity
      tags: [auth]
      summary: Begin linking an external identity (OIDC purpose=link).
      description: |
        An account-security mutation: verifies the pre-existing password up
        front and starts an OIDC link transaction bound to the acting session.
        Returns the IdP authorization URL; the link lands at the callback.
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityLinkRequest"
      responses:
        "200":
          description: The IdP authorization URL for the link.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcStartResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/auth/identities/{id}:
    parameters:
      - $ref: "#/components/parameters/IdentityID"
    delete:
      operationId: unlinkIdentity
      tags: [auth]
      summary: Unlink an external identity.
      description: |
        An account-security mutation proved by the pre-existing password. Refuses
        removing the last remaining credential; deletes the identity's sessions;
        reissues the acting session. The reissued token replaces the old one.
      x-hikyo-class: unauthenticated
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdentityUnlinkRequest"
      responses:
        "200":
          description: Unlinked. The reissued session token replaces the old one.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
...
  /api/v1/instance/oidc-providers:
    get:
      operationId: listOidcProviders
      tags: [instance]
      summary: List configured OIDC providers.
      x-hikyo-class: instance
      x-hikyo-operation: oidc-provider.list
      x-hikyo-formula: [instance-config@instance]
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      responses:
        "200":
          description: The configured providers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcProviderList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/instance/oidc-providers/{slug}:
    parameters:
      - $ref: "#/components/parameters/ProviderSlugPath"
    get:
      operationId: getOidcProvider
      tags: [instance]
      summary: Read one OIDC provider.
      x-hikyo-class: instance
      x-hikyo-operation: oidc-provider.get
      x-hikyo-formula: [instance-config@instance]
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      responses:
        "200":
          description: The provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcProvider"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
    put:
      operationId: putOidcProvider
      tags: [instance]
      summary: Create or reconfigure an OIDC provider.
      description: |
        The issuer is immutable after create (A3): a changed issuer is refused.
        Discovery is re-run and the document's issuer must byte-equal the
        configured one. A reconfigure sweeps every session authenticated through
        the provider (A4).
      x-hikyo-class: instance
      x-hikyo-operation: oidc-provider.put
      x-hikyo-formula: [instance-config@instance]
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OidcProviderInput"
      responses:
        "200":
          description: The created or reconfigured provider.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OidcProvider"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
    delete:
      operationId: deleteOidcProvider
      tags: [instance]
      summary: Delete an OIDC provider.
      description: |
        Sweeps every session authenticated through the provider (A4); its live
        transactions cascade (A14).
      x-hikyo-class: instance
      x-hikyo-operation: oidc-provider.delete
      x-hikyo-formula: [instance-config@instance]
      x-hikyo-artifacts: [human-session]
      x-hikyo-min-revision: 1
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    hikyoSession:
      type: http
      scheme: bearer
      bearerFormat: hik_1_cli_...
      description: |
        A CLI session artifact minted by an identity-protocol endpoint. The
        server trusts nothing inside the value: kind, principal, expiry and
        credential epoch are read from the row the verifier resolves to.
  parameters:
...
    ProviderSlug:
      name: provider
      in: path
      required: true
      description: Identity-provider slug.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-z0-9][a-z0-9-]{0,63}$"
    ProviderSlugPath:
      name: slug
      in: path
      required: true
      description: Identity-provider slug.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-z0-9][a-z0-9-]{0,63}$"
    IdentityID:
      name: id
      in: path
      required: true
      description: External identity identifier.
      schema:
        $ref: "#/components/schemas/ID"
...
  responses:
    BadRequest:
      description: |
        The request does not satisfy this document. Decided before any tenant
        resolution, so `detail` leaks nothing about tenancy — it is the only
        error response permitted to carry one.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: |
        No usable authentication artifact was presented. Uniform: absent,
        malformed, unknown, expired, revoked and epoch-superseded artifacts
        are indistinguishable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: |
        Either the principal does not hold the operation's formula at instance
        scope — instance-class operations have no tenant object whose
        nonexistence could be mimicked, so the probe contract there is grant
        refusal, not tenancy — or the principal DOES hold it and the acting
        session's assurance is inadequate for an MFA-mandatory operation.

        The second case is why two tenant-scoped operations (`renameOrg`,
        `deleteOrg`) declare this status: their formula atom `instance-config`
        is MFA-mandatory, and the refusal fires only AFTER the grant check
        succeeded. A caller who reaches it can already reach the object, so
        naming the step-up discloses nothing the uniform 404 was protecting —
        and hiding it would tell a capability holder the object is missing.
        Grant refusal on a tenant-scoped operation is always the 404.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: |
        The addressed object does not exist **or** the principal may not reach
        it — indistinguishable by design, byte-identical in status and body.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    TooManyRequests:
      description: |
        The instance-wide admission budget or a per-source limit is
        exhausted. Uniform on every path, with no unbounded work performed.
      headers:
        Retry-After:
          required: true
          description: Whole seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Internal:
      description: An unexpected server fault. The cause is logged, never returned.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ID:
      type: string
      description: A prefixed UUIDv7, e.g. `org_0198…`.
      minLength: 3
      maxLength: 64
      pattern: "^[a-z]{2,8}_[0-9a-fA-F-]{36}$"
...
    LoginResult:
      type: object
      required: [session, principal]
      properties:
        session_token:
          type: string
          description: |
            The bearer value, returned exactly once. A replayable credential:
            cookie attributes constrain browsers, not an attacker holding the
            value, which is why lifetimes are short and revocation is a
            delete. OMITTED for a browser-artifact session: that token is
            delivered ONLY on the `__Host-hikyo` HttpOnly cookie so injected
            same-origin script cannot read it. A CLI-artifact session, which has
            no cookie channel, still carries it here.
        session:
          $ref: "#/components/schemas/Session"
        principal:
          $ref: "#/components/schemas/Principal"
...
    AuthMethods:
      type: object
      additionalProperties: false
      required: [providers, local_login_enabled]
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/AuthMethodProvider"
        local_login_enabled:
          type: boolean
...
    OidcStartRequest:
      type: object
      additionalProperties: false
      required: [purpose]
      properties:
        purpose:
          type: string
          enum: [login, link, reauth]
        environment_id:
          type: string
          maxLength: 64
          description: Required for reauth; the window scope.
        proof:
          type: string
          maxLength: 1024
          description: Required for link; the pre-existing password.
    OidcStartResult:
      type: object
      additionalProperties: false
      required: [authorization_url]
      properties:
        authorization_url:
          type: string
          description: The IdP authorization URL to redirect the browser to.
...
    IdentityLinkRequest:
      type: object
      additionalProperties: false
      required: [provider, proof]
      properties:
        provider:
          type: string
          maxLength: 64
        proof:
          type: string
          maxLength: 1024
    IdentityUnlinkRequest:
      type: object
      additionalProperties: false
      required: [proof]
      properties:
        proof:
          type: string
          maxLength: 1024
...
    IdentityList:
      type: object
      additionalProperties: false
      required: [identities]
      properties:
        identities:
          type: array
          items:
            $ref: "#/components/schemas/ExternalIdentity"
...
    OidcProviderList:
      type: object
      additionalProperties: false
      required: [providers]
      properties:
        providers:
          type: array
          items:
            $ref: "#/components/schemas/OidcProvider"
...
security:
  - hikyoSession: []